Replace the simple planet example with the full one, renamed Planet Example
The old PlanetExample was a reduced version from before the JSX layer
supported the full feature set. FullPlanetExample is renamed to
PlanetExample and takes its slot; files renamed to match their exports
(InteractivePlanetExample gets its own file).
